READ COMMENT >< - Updated point plotting but..
Noticed some scaling issues with the y. The y scale doesn't go all the
way up to the right value. It's just that the labels for the y axis on
the side are wrong and the point plotting is right. You can kind of
debug it and watch my yMax and yMin values to see what I'm talking
about. I think the problem is that the yValString  length can be longer
than 10 so that when you're doing the drawString, you don't get up to
the highest values in that array.
Example:
y = x+5 w/ increments of 1 starting at 0
(y axis should go up to 15 but only goes up to 10)